Move dao methods and documentation into AbstractDatabase
